Rename domain interfaces Context and Given
In order to represet the domain better we rename two domain interfaces:

Context to Criteria (_noun, a principle or standard by which
something may be judged or decided_).

Given to Context (_noun, the circumstances that form the setting for
an event, statement, or idea, and in terms of which it can be fully
understood and assessed_).