Add server with express
In order to do so we create a decoupled app module with the purpose of
setup routes and middlewares. The app module exports a function called
buildApp and we mostly use it to create dummy applications in our tests
due to the fact we need to ensure that everything is running smoothly
within fast feedback test suites.