Add onError middleware
We add onError middleware to centralize the API error handling in one
single point (if possible).

In order to map domain errors we create an ApiError class to hold
standard HTTP information such as status code and a human readable code.

Throw an error from the depths of the application must be treated in top
level API layer, for instances routers.