Create memory store module
Move the storage resposability from the repository to one deep layer in
order to map needed persistance and fetching features for the feature.
This also delegates the decision of a main data storage for the future
unblocking us to keep working in features right now.

One of the goals that we're trying to achieve in easy-toogles is the
fast and simple configuration for multiple types of environments which
includes different kind of data storage tools.