Add destroy function in the memory store
In order to provide a delete capability for Application we add a new
function in the memory store and map the Store interface as well.

A change in the memory store was introduced in order to create an
isolation between tests. To achieve that we create two functions to
mutate directely the variable documents (storage) inside the store (use
it wisely).