Add accessKeyId and secretAccessKey for the Dynamo config
Inside Docker the AWS SDK is requesting the access key ID and secret
access key which is making the dynamo client test fails in a container
consequentely in Travis CI as well. In order to fix this issue we're
including the accessKeyId and secretAccessKey in the client config
interface to setup it programmaticaly before the module execute.

Also remove the object-literal-sort-keys configuration from TSLint due
to the fact some attributes are better organized close to each other in
a logical way, for instance, credentials like access keys and secrets.