Add Apollo and GraphQL minimal setup
Add a GraphQL query to find an Application by ID.

Also, add GraphQL custom scalar type JSON in order to keep the
Application interface of criteria whereas the keys are dynamic. This
practice is discouraged by the community since your schema should be
strongly typed and predictable to consumers/clients. However, we did it
anyway to avoid change the toggles evaluation core and existing RESTful
endpoints. In the future, we'll have to update the schema to be more
GraphQL way.

See below the actual Application contract and a suggestion.

Actual:

```json
{
   "features" : [
      {
         "name" : "foo",
         "criteria" : {
            "country" : ["BR", "CL"]
         }
      },
   ],
   "name" : "FooApp",
   "id" : "fooapp-fake-id"
}
```

Suggestion:

```json
{
   "features" : [
      {
        "name": "foo",
        "criterias": [
          {
            "name": "country",
            "values": ["BR", "CL"]
          }
        ]
      }
   ],
   "name" : "FooApp",
   "id" : "fooapp-fake-id"
}
```

Change the tsconfig file to include src directory, skip lib check, and
libs es6 and esnext.asynciterable in order to make Apollo and
GraphQL work.

Add new task `start:dev` which uses nodemon under the hood to refresh
the server after changes.