Fix start:prod file
In the tsconfig.json we're only including the src/ directory. This
configuration is an instructing of TypeScript compiler to only compile
files inside this folder into dist.

Surprisingly this error wasn't noticed before probably due to we are not
cleaning the dist folder before building. This issue was also fixed
including a `rm -rf dist/` before running tsc command.