Add Prettier
Code standard is important and in order to keep consistency in the
project style guides we did install Prettier with a basic setup close to
our tslint configuration. We also added tslint-config-prettier to
include Prettier rules in the lint verification as well.

Two new tasks were included in the package.json: format and
format:check.

- format should apply prettier changes in all .ts files inside src and
  ts directories.
- format:check should verify possible incosistencies in all .ts files
  inside src and test directories.