Move Graph domain to a separate directory
In order to keep Graph domain separate we create a new directory called
graph. The typeDefs are in a separate module graph/schema as well.

Refactor resolvers to access the store from the context instead of curry
function with dependency injection. This approach is used in Apollo
applications to share data between resolvers.

Reference:
- https://www.apollographql.com/docs/apollo-server/essentials/data.html#context