Add a query to evaluate toggles
This commit introduces the Toggle type and the toggle query in the graph
schema, both are needed to the toggles evaluation. Also, add a new
resolver to transform the query input context into the domain context
interface. The resolver will responsible for transform the input context
into the domain context but this might change in the future if the context
interface doesn't have dynamic key names (this might impact the toggle
rest endpoint instead of the graph resolver).

```
query evaluate {
  toggle(
    applicationId: "APPLICATION_ID"
    context: [{ name: "country", value: "AU" }]
  ) {
    name
    isActive
  }
}

{
  "data": {
    "toggle": [
      {
        "name": "foo",
        "isActive": true
      }
    ]
  }
}
```

Increase the type safety of the toggle interface by changing the dynamic
key name for a known key and value. A possible improvement for the
future is maybe to apply the same mindset in the domain context
interface as well.

Move graph server integration tests to a separate directory and also
split them in two, one for queries and other for mutations in order to
improve the discoverability.