Add ID key to Application, Feature and Criteria
At the moment, the application doesn't support partial updates, for
instance, change a single criteria inside an application. This commit
add ID keys to Application, Feature, and Criteria as a first step to
provide partial updates in the future.

In the future the plan is support multiple types of stores. That being
said, the responsability of generate and add ID keys to objects before
persistence was moved from the store to the repository in order to keep
consistency among all store types and avoid code duplication.